feat: fold the minroot_serde example under a feature
- Removed `examples/minroot_serde.rs` file, which previously managed the serialization and deserialization of public params demonstrated on the Nova VDF,
- Created and updated several utility functions and data structures in `examples/minroot.rs` to accommodate the prior `minroot_serde` demonstration.
- Introduced the `abomonate` feature,
