+++ Cargo.toml
@@ -82,6 +82,7 @@ harness = false
 
 [features]
 default = []
+abomonate = []
 asm = ["halo2curves/asm"]
 # Compiles in portable mode, w/o ISA extensions => binary can be executed on all systems.
 portable = ["pasta-msm/portable"]
